pyproject.toml
[project]
name = "wcpan-drive-cli"
version = "3.9.0"
description = "command line tool for wcpan.drive"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Wei-Cheng Pan", email = "legnaleurc@gmail.com"}
]
requires-python = ">=3.12,<4.0"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Typing :: Typed",
]
dependencies = [
    "PyYAML>=6.0.1,<7.0.0",
    "wcpan-drive-core>=4.1.0,<5.0.0",
    "wcpan-logging>=2.0.3,<3.0.0",
    "wcpan-queue>=7.2.1,<8.0.0",
    "python-magic>=0.4.27,<0.5.0",
    "pymediainfo>=7.0.1,<8.0.0",
]

[project.scripts]
"wcpan.drive.cli" = "wcpan.drive.cli._main:main"

[dependency-groups]
dev = [
    "ruff>=0.11.2,<0.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/wcpan"]
